requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
torch==2.1.1
torch-geometric==2.3.1
neo4j==5.14.1
supabase==2.0.3
python-dotenv==1.0.0
pydantic==2.5.2
networkx==3.2.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
confluent-kafka==2.3.0
avro-python3==1.10.2 